move store from share folder
